Доработана система менеджера API
