+++ src/content/docs/index.mdx
@@ -7,7 +7,7 @@ hero:
     file: ../../assets/houston.webp
   actions:
     - text: Посмотреть РМы
-      link: /chat-phrases
+link: /chat-phrasesBack
       icon: right-arrow
     - text: Установить расширение
       link: https://helper.chrsnv.ru
@@ -21,7 +21,7 @@ prev: false
 next: false
 ---
 
-import { Card, CardGrid } from "@astrojs/starlight/components";
+import {Card, CardGrid} from "@astrojs/starlight/components";
 
 ## Возможности
