+++ src/pages/admin.astro
@@ -8,7 +8,7 @@ import AdminPanel from '../components/AdminPanel';
     <link rel="icon" type="image/svg+xml" href="/favicon.svg"/>
     <meta name="viewport" content="width=device-width"/>
     <meta name="generator" content={Astro.generator}/>
-    <title>Admin Panel</title>
+    <title>Панель админа</title>
 </head>
 <body>
 <AdminPanel client:load/>
